+++ lib/src/languages/dart/dart_grammar.dart
@@ -460,9 +460,7 @@ class DartGrammarDefinition extends DartGrammarLexer {
       });
 
   Parser<ASTExpression?> switchCaseLabel() =>
-      ((string('case').trimHidden() &
-                      ref0(expression) &
-                      char(':').trimHidden())
+      ((string('case').trimHidden() & ref0(expression) & char(':').trimHidden())
                   .map((v) => v[1] as ASTExpression?) |
               (string('default').trimHidden() & char(':').trimHidden()).map(
                 (v) => null,