+++ lib/src/languages/typescript/ts/typescript_grammar.dart
@@ -487,9 +487,7 @@ class TypeScriptGrammarDefinition extends TypeScriptGrammarLexer {
       });
 
   Parser<ASTExpression?> switchCaseLabel() =>
-      ((string('case').trimHidden() &
-                      ref0(expression) &
-                      char(':').trimHidden())
+      ((string('case').trimHidden() & ref0(expression) & char(':').trimHidden())
                   .map((v) => v[1] as ASTExpression?) |
               (string('default').trimHidden() & char(':').trimHidden()).map(
                 (v) => null,